版本 1.3.0

See merge request xugu-open/dbeaver!2

+++ plugins/org.jkiss.dbeaver.ext.postgresql/src/org/jkiss/dbeaver/ext/postgresql/edit/PostgreMViewManager.java
@@ -1,6 +1,6 @@
 /*
  * DBeaver - Universal Database Manager
- * Copyright (C) 2010-2020 DBeaver Corp and others
+ * Copyright (C) 2010-2021 DBeaver Corp and others
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,13 +21,16 @@ import org.jkiss.dbeaver.ext.postgresql.model.PostgreMaterializedView;
 import org.jkiss.dbeaver.ext.postgresql.model.PostgreSchema;
 import org.jkiss.dbeaver.ext.postgresql.model.PostgreTablespace;
 import org.jkiss.dbeaver.ext.postgresql.model.PostgreViewBase;
+import org.jkiss.dbeaver.model.DBConstants;
 import org.jkiss.dbeaver.model.DBPEvaluationContext;
 import org.jkiss.dbeaver.model.DBUtils;
 import org.jkiss.dbeaver.model.edit.DBECommandContext;
 import org.jkiss.dbeaver.model.edit.DBEPersistAction;
 import org.jkiss.dbeaver.model.exec.DBCExecutionContext;
 import org.jkiss.dbeaver.model.impl.edit.SQLDatabasePersistAction;
+import org.jkiss.dbeaver.model.impl.sql.edit.struct.SQLTableManager;
 import org.jkiss.dbeaver.model.runtime.DBRProgressMonitor;
+import org.jkiss.utils.ArrayUtils;
 
 import java.util.List;
 import java.util.Map;
@@ -43,18 +46,28 @@ public class PostgreMViewManager extends PostgreViewManager {
         PostgreSchema schema = (PostgreSchema)container;
         PostgreMaterializedView newMV = new PostgreMaterializedView(schema);
         newMV.setName("new_mview"); //$NON-NLS-1$
+        setNewObjectName(monitor, schema, newMV);
         return newMV;
     }
 
     @Override
-    protected void createOrReplaceViewQuery(DBRProgressMonitor monitor, List<DBEPersistAction> actions, PostgreViewBase view) throws DBException {
-        super.createOrReplaceViewQuery(monitor, actions, view);
+    protected String getBaseObjectName() {
+        return SQLTableManager.BASE_MATERIALIZED_VIEW_NAME;
+    }
+
+    @Override
+    protected void createOrReplaceViewQuery(DBRProgressMonitor monitor, List<DBEPersistAction> actions, PostgreViewBase view, Map<String, Object> options) throws DBException {
+        super.createOrReplaceViewQuery(monitor, actions, view, options);
         // Indexes DDL
     }
 
     @Override
     public void appendViewDeclarationPrefix(DBRProgressMonitor monitor, StringBuilder sqlBuf, PostgreViewBase view) throws DBException {
-        PostgreMaterializedView mview = (PostgreMaterializedView)view;
+        PostgreMaterializedView mview = (PostgreMaterializedView) view;
+        String[] relOptions = mview.getRelOptions();
+        if (!ArrayUtils.isEmpty(relOptions)) {
+            sqlBuf.append("\nWITH(").append(String.join("," , relOptions)).append(")");
+        }
         PostgreTablespace tablespace = mview.getTablespace(monitor);
         if (tablespace  != null) {
             sqlBuf.append("\nTABLESPACE ").append(DBUtils.getQuotedIdentifier(tablespace));
@@ -71,12 +84,12 @@ public class PostgreMViewManager extends PostgreViewManager {
     @Override
     protected void addObjectModifyActions(DBRProgressMonitor monitor, DBCExecutionContext executionContext, List<DBEPersistAction> actionList, ObjectChangeCommand command, Map<String, Object> options) throws DBException {
         final PostgreMaterializedView mView = (PostgreMaterializedView) command.getObject();
-        super.addObjectDeleteActions(monitor, executionContext, actionList, new ObjectDeleteCommand(mView, "Drop view"), options);
-        super.addObjectModifyActions(monitor, executionContext, actionList, command, options);
-
+        if (!command.hasProperty(DBConstants.PROP_ID_DESCRIPTION) || command.getProperties().size() > 1) {
+            super.addObjectDeleteActions(monitor, executionContext, actionList, new ObjectDeleteCommand(mView, "Drop view"), options);
+            super.addObjectModifyActions(monitor, executionContext, actionList, command, options);
+        }
         if (command.hasProperty("tablespace")) {
-            final String alterPrefix = "ALTER " + mView.getViewType() + " " + command.getObject().getFullyQualifiedName(DBPEvaluationContext.DDL) + " ";
-
+            final String alterPrefix = "ALTER " + mView.getTableTypeName() + " " + command.getObject().getFullyQualifiedName(DBPEvaluationContext.DDL) + " ";
             actionList.add(new SQLDatabasePersistAction(alterPrefix + "SET TABLESPACE " + mView.getTablespace(monitor).getName()));
         }
     }
